Restructure library docs
This commit creates a new top-level section, "Component library", which
contains documentation on how to contribute components to the library.

This splits it out from the existing Moodle section, which focuses
primarily on consuming the existing content of the library.

This helps to distinguish between the two main user groups:
- those generating new components; and
- those primarily consuming existing components.

Fixes #4